+++ scripts/distributables/linux-build-distributable-concordium-client.sh
@@ -1,9 +1,12 @@
 #!/usr/bin/env sh
 
+set -ex
+
 cd /build
 
 sed -i "s/default: False/default: True/g" deps/concordium-base/package.yaml
 
 stack build --stack-yaml stack.linux-static.yaml
 
-cp $(stack --stack-yaml stack.linux-static.yaml path --local-install-root)/bin/concordium-client /out/concordium-client
+LOCAL_INSTALL_ROOT=$(stack --stack-yaml stack.linux-static.yaml path --local-install-root)
+cp "$LOCAL_INSTALL_ROOT"/bin/concordium-client /out/concordium-client